+++ components/support/nimbus-fml/src/backends/frontend_manifest.rs
@@ -143,6 +143,7 @@ impl From<PropDef> for FeatureFieldBody {
     fn from(value: PropDef) -> Self {
         Self {
             pref_key: value.pref_key.clone(),
+            gecko_pref: value.gecko_pref.clone(),
             string_alias: value.string_alias.as_ref().map(TypeRef::to_string),
             field: value.into(),
         }