+++ components/support/nimbus-fml/src/backends/kotlin/templates/FeatureManifestTemplate.kt
@@ -104,6 +104,23 @@ object {{ nimbus_object }} : FeatureManifestInterface<{{ nimbus_object }}.Featur
             {{- f|quoted }}
             {%- if !loop.last %}, {% endif %}
             {%- endfor %})
+
+    /**
+     * Get a map of prefs to their repective feature IDs.
+     */
+    override fun geckoPrefsMap(): Map<String, Map<String, GeckoPref>> =
+        mapOf(
+            {%- for f in self.fm.iter_features_with_prefs() %}
+            {{ f.0|quoted }} to mapOf(
+                {%- for p in f.1.iter() %}
+                {{ p.0|quoted }} to GeckoPref({{ p.1.pref()|quoted }}, {{ p.1.branch()|quoted }})
+                {%- if !loop.last %},{% endif -%}
+                {% endfor %}
+            )
+            {%- if !loop.last %},{% endif -%}
+            {% endfor %}
+        )
+
     /**
      * Accessor object for generated configuration classes extracted from Nimbus, with built-in
      * default values.