+++ components/support/nimbus-fml/src/fixtures/mod.rs
@@ -6,7 +6,9 @@ use std::collections::BTreeMap;
 
 use serde_json::Value;
 
-use crate::intermediate_representation::{EnumDef, ObjectDef, PropDef, TypeRef, VariantDef};
+use crate::intermediate_representation::{
+    EnumDef, GeckoPrefDef, ObjectDef, PrefBranch, PropDef, TypeRef, VariantDef,
+};
 
 pub(crate) mod intermediate_representation;
 
@@ -53,6 +55,28 @@ impl PropDef {
             default: value.clone(),
             doc: format!("{nm} property of type {typ}"),
             pref_key: None,
+            gecko_pref: None,
+            string_alias: None,
+        }
+    }
+
+    pub(crate) fn new_with_gecko_pref(
+        nm: &str,
+        typ: &TypeRef,
+        value: &Value,
+        pref_key: &str,
+        pref_branch: PrefBranch,
+    ) -> Self {
+        Self {
+            name: nm.to_string(),
+            typ: typ.clone(),
+            default: value.clone(),
+            doc: format!("{nm} property of type {typ}"),
+            pref_key: None,
+            gecko_pref: Some(GeckoPrefDef {
+                pref: pref_key.into(),
+                branch: pref_branch,
+            }),
             string_alias: None,
         }
     }
@@ -64,6 +88,7 @@ impl PropDef {
             default: value.clone(),
             doc: nm.to_string(),
             pref_key: None,
+            gecko_pref: None,
             string_alias: Some(sa.clone()),
         }
     }
@@ -75,6 +100,7 @@ impl PropDef {
             typ: typ.clone(),
             default: default.clone(),
             pref_key: None,
+            gecko_pref: None,
             string_alias: None,
         }
     }