+++ components/support/nimbus-fml/src/frontend.rs
@@ -13,8 +13,8 @@ use crate::{
     defaults::DefaultsMerger,
     error::Result,
     intermediate_representation::{
-        EnumDef, FeatureDef, FeatureManifest, ModuleId, ObjectDef, PropDef, TargetLanguage,
-        TypeRef, VariantDef,
+        EnumDef, FeatureDef, FeatureManifest, GeckoPrefDef, ModuleId, ObjectDef, PropDef,
+        TargetLanguage, TypeRef, VariantDef,
     },
     parser::get_typeref_from_string,
 };
@@ -42,6 +42,10 @@ pub(crate) struct FeatureFieldBody {
     #[serde(default)]
     #[serde(skip_serializing_if = "Option::is_none")]
     pub(crate) pref_key: Option<String>,
+
+    #[serde(default)]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub(crate) gecko_pref: Option<GeckoPrefDef>,
 
     #[serde(default)]
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -346,6 +350,7 @@ impl ManifestFrontEnd {
     fn get_prop_def_from_feature_field(&self, nm: &str, body: &FeatureFieldBody) -> PropDef {
         let mut prop = self.get_prop_def_from_field(nm, &body.field);
         prop.pref_key.clone_from(&body.pref_key);
+        prop.gecko_pref.clone_from(&body.gecko_pref);
         if let Some(s) = &body.string_alias {
             prop.string_alias = Some(TypeRef::StringAlias(s.clone()));
         }
@@ -379,8 +384,9 @@ impl ManifestFrontEnd {
                 }
             },
             default: json!(body.default),
-            pref_key: None,
             string_alias: None,
+            pref_key: Default::default(),
+            gecko_pref: Default::default(),
         }
     }