+++ megazords/ios-rust/Sources/MozillaRustComponentsWrapper/Nimbus/FeatureManifestInterface.swift
@@ -37,4 +37,6 @@ public protocol FeatureManifestInterface {
     func getFeature(featureId: String) -> FeatureHolderAny?
 
     func getCoenrollingFeatureIds() -> [String]
+
+    func geckoPrefsMap() -> [String: [String: String]]
 }